# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Viral Vertical Videos Creator API",
    "version": "2.0.0",
    "description": "Especificação da API REST Serverless para a plataforma Web SaaS de automação de vídeos verticais."
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Servidor Local de Desenvolvimento"
    },
    {
      "url": "https://viralcreator.app/api",
      "description": "Servidor de Produção (Vercel)"
    }
  ],
  "paths": {
    "/trends/search": {
      "post": {
        "summary": "Busca tendências virais com Virality Score e cache Redis",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "tema": { "type": "string", "example": "GTA 6" },
                  "source": { "type": "string", "example": "all" },
                  "idioma": { "type": "string", "example": "pt" },
                  "pais": { "type": "string", "example": "BR" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Lista de tendências retornada com sucesso" }
        }
      }
    },
...
  }
}